feat: add various structures for GitHub workflow management including artifacts, environment, defaults, and strategies
