+++ crates/gh-workflow/src/permissions.rs
